Add test cases for multiple uninitialised variable declarations
